+++ src/frontend/commands/CrawlController.php
@@ -26,30 +26,26 @@ class CrawlController extends \luya\console\Command
 {
     public function actionIndex()
     {
-        try {
-            // sart time measuremnt
-            $start = microtime(true);
-            
-            $container = new CrawlContainer([
-                'baseUrl' => $this->module->baseUrl,
-                'filterRegex' => $this->module->filterRegex,
-                'verbose' => $this->verbose,
-                'doNotFollowExtensions' => $this->module->doNotFollowExtensions,
-                'useH1' => $this->module->useH1,
-            ]);
-            
-            $timeElapsed = round((microtime(true) - $start) / 60, 2);
-            
-            $table = new Table();
-            $table->setHeaders(['status', 'url', 'message']);
-            $table->setRows($container->getReport());
-            $this->output($table->run());
-            $this->outputInfo('memory usage: ' . FileHelper::humanReadableFilesize(memory_get_usage()));
-            $this->outputInfo('memory peak usage: ' . FileHelper::humanReadableFilesize(memory_get_peak_usage()));
-            
-            return $this->outputSuccess('Crawler finished in ' . $timeElapsed . ' min.');
-        } catch (\Exception $e) {
-            return $this->outputError('Exception in file "'.$e->getFile() . '" on line #' . $e->getLine() . ': ' . $e->getMessage());
-        }
+        // sart time measuremnt
+        $start = microtime(true);
+        
+        $container = new CrawlContainer([
+            'baseUrl' => $this->module->baseUrl,
+            'filterRegex' => $this->module->filterRegex,
+            'verbose' => $this->verbose,
+            'doNotFollowExtensions' => $this->module->doNotFollowExtensions,
+            'useH1' => $this->module->useH1,
+        ]);
+        
+        $timeElapsed = round((microtime(true) - $start) / 60, 2);
+        
+        $table = new Table();
+        $table->setHeaders(['status', 'url', 'message']);
+        $table->setRows($container->getReport());
+        $this->output($table->run());
+        $this->outputInfo('memory usage: ' . FileHelper::humanReadableFilesize(memory_get_usage()));
+        $this->outputInfo('memory peak usage: ' . FileHelper::humanReadableFilesize(memory_get_peak_usage()));
+        
+        return $this->outputSuccess('Crawler finished in ' . $timeElapsed . ' min.');
     }
 }