+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file. This projec
 ## 1.0.2 (in progress)
 
 + [#5](https://github.com/luyadev/luya-module-crawler/issues/5) Add option to provide group search in default controller.
++ [#4](https://github.com/luyadev/luya-module-crawler/issues/4) Add info when base url does not return status code 200.
 
 ## 1.0.1 (28. March 2018)